Add DbContext Configuration for ProductDbContext
This commit introduces using directives for `eCommerce.Data and `Microsoft.EntityFrameworkCore`. It configures the `ProductDbContext` to use SQL Server with a connection string named `DefaultConnection`. Additionally, it initializes the wb application builder and builds the application instance setting up the foundation for database operations in the ASP.NET Core application.