Change Price type to deciam and update migrations
Updatedt the 'Price' property in the 'Product' class from the 'double' to 'decimal' for better precision. Created initial migration to set up the 'Products' table with a 'Price' column of type 'float', followed by a second migration to alter the 'Price' column to 'decimal(18.2)'. Updated related designer files and the model snapshot to reflect these changes.