Issue#9 create db context (#16)
* Add Entity Framework Core packages to project

Updated `eCommerce.csproj` to include references for `Microsoft.EntityFrameworkCore.SqlServer` and Microsoft.EntityFrameworkCore.Tools` (version 9.0.7).
Configured `PrivateAssets` and `IncludeAssets` for `Microsoft.EntityFrameworkCore.Tools` to manage usage.

* Add ProductDbContext for managing Product entities

Introduces the `ProductDbContext` class inheriting from `DbContext to handle the database context for `Product` entities. Includes a constructor for `DbContextOptions` and a `DbSet<Product>` property to track `Product` instances.

* Add connection string for local SQL Server database

Added a `ConnectionStrings` section in appsettings.json with a `DefaultConnection` string for connecting to the GameStoreDb databse using local SQL Server.

* Add DbContext Configuration for ProductDbContext

This commit introduces using directives for `eCommerce.Data and `Microsoft.EntityFrameworkCore`. It configures the `ProductDbContext` to use SQL Server with a connection string named `DefaultConnection`. Additionally, it initializes the wb application builder and builds the application instance setting up the foundation for database operations in the ASP.NET Core application.

* Added a markdown file for coding resources

* Change Price type to deciam and update migrations

Updatedt the 'Price' property in the 'Product' class from the 'double' to 'decimal' for better precision. Created initial migration to set up the 'Products' table with a 'Price' column of type 'float', followed by a second migration to alter the 'Price' column to 'decimal(18.2)'. Updated related designer files and the model snapshot to reflect these changes.