Create product listing page (#18)
* Add `Product Listings` navigation item

Introduced a new navigation link for `Product Listings` in
the `_Layout.cshtml` file, linking to the `Index` action of the `Product` controller.

* Enhance product listing functionality

- Made the `Index` action in `ProductController` asynchronous to fetch products from the database.
- Updated the `Price` Property in the `Product` class to include a currency data type.
- Improved the `Index.cshtml` view to display a list of products, including a title, a create button, and a table for product details.

* Add success message for product creation

Updated `ProductController` to store a success message in `TempData` after a product is created.
Modified `Index.cshtml` to display this message in a Bootstrap alert box, enhancing user feedback.